// src/notific8/notific8-options.d.ts
export interface Notific8Options {
  closeText?: string;
  horizontalEdge?: 'top'|'bottom';
  life?: number;
  queue?: boolean;
  sticky?: boolean;
  styleNamespace?: string;
  theme?: string;
  themeColor?: string;
  verticalEdge?: 'left'|'right';
  zIndex?: number;
  [key: string]: string|number|boolean|undefined;
}